Update RegexBlock for MW 1.35+/the new blocking system
This is ugly, messy and generally speaking just awful...but it seemed to work?
The new RegularExpressionDatabaseBlock class should probably supersede some of the older methods, of which at least one is now unused.
However, as the inline comments demonstrate, this is not yet possible for all the methods as some are buggy and thus we have basically two copies of similar-ish methods (newer one based on the core Block class(es) and older, custom RegexBlock ones).

Bug: T236058
Bug: T243702
Bug: T248060
Change-Id: I74a2875f59cb4e47b64dae4df2584920f4a3640c
(cherry picked from commit c45085d00a4621a20bb5407cb3820786e02e33a1)